=removing support for loading certs via file paths, adding host factory module, removing config section
